changed integers to floats in muigration data

+++ Giggity/db/migrate/20191203224942_create_agency_gigs.rb
@@ -2,7 +2,7 @@ class CreateAgencyGigs < ActiveRecord::Migration[6.0]
   def change
     create_table :agency_gigs do |t|
       t.string :date
-      t.integer :compensation
+      t.float :compensation
 
       t.timestamps
     end
